+++ crates/next-api/src/nft.rs
@@ -3,27 +3,30 @@ use std::{
     ops::Deref,
 };
 
-use anyhow::{Context, Result};
+use anyhow::{Context, Result, bail};
 use bincode::{Decode, Encode};
 use next_core::{app_structure::FileSystemPathVec, next_config::NextConfig};
+use rustc_hash::{FxHashMap, FxHashSet};
 use tracing::Instrument;
 use turbo_rcstr::RcStr;
 use turbo_tasks::{
-    FxIndexMap, NonLocalValue, ResolvedVc, TryFlatJoinIterExt, TryJoinIterExt, Vc,
-    trace::TraceRawVcs,
+    FxIndexMap, FxIndexSet, NonLocalValue, ReadRef, ResolvedVc, TryFlatJoinIterExt, TryJoinIterExt,
+    Vc, trace::TraceRawVcs,
 };
 use turbo_tasks_fs::{
     DirectoryEntry, FileSystemPath,
     glob::{Glob, GlobOptions},
 };
-use turbopack_core::{module::Module, module_graph::ModuleGraph};
-
-use crate::{
-    nft_json::{
-        TracedModuleData, relativize_glob, traced_module_data_for_graph, traced_modules_for_entries,
-    },
-    project::Project,
+use turbo_tasks_hash::HashAlgorithm;
+use turbopack_core::{
+    asset::Asset,
+    chunk::{ChunkingType, TracedMode},
+    ident::AssetIdent,
+    module::{Module, Modules},
+    module_graph::{GraphTraversalAction, ModuleGraph},
 };
+
+use crate::project::Project;
 
 #[turbo_tasks::value]
 pub struct EndpointTraceResult {
@@ -316,4 +319,354 @@ pub async fn tracing_exclude_glob(
     } else {
         Vc::cell(None)
     })
+}
+
+#[turbo_tasks::function]
+pub async fn traced_modules_for_entries(
+    module_graph: Vc<ModuleGraph>,
+    entry_modules: Vc<Modules>,
+    exclude_glob: Option<Vc<Glob>>,
+    entries_are_traced: bool,
+) -> Result<Vc<Modules>> {
+    let exclude_glob = if let Some(exclude_glob) = exclude_glob {
+        Some(exclude_glob.await?)
+    } else {
+        None
+    };
+    let module_idents = if exclude_glob.is_some() {
+        let data = traced_module_data_for_graph(module_graph, entries_are_traced).await?;
+        Some(data.idents.await?)
+    } else {
+        None
+    };
+
+    let mut traced_modules = FxIndexSet::default();
+    module_graph.await?.traverse_edges_dfs(
+        entry_modules.await?.iter().copied(),
+        &mut (),
+        |parent, target, _| {
+            let Some((parent, ref_data)) = parent else {
+                if entries_are_traced {
+                    traced_modules.insert(target);
+                }
+                return Ok(GraphTraversalAction::Continue);
+            };
+
+            if should_visit_for_tracing(&ref_data.chunking_type, traced_modules.contains(&parent)) {
+                if let Some(exclude_glob) = &exclude_glob
+                    && exclude_glob.matches(
+                        &module_idents
+                            .as_ref()
+                            .unwrap()
+                            .get(&target)
+                            .context("missing path for module")?
+                            .path
+                            .path,
+                    )
+                {
+                    return Ok(GraphTraversalAction::Skip);
+                }
+                traced_modules.insert(target);
+            };
+            Ok(GraphTraversalAction::Continue)
+        },
+        |_, _, _| Ok(()),
+        true,
+    )?;
+
+    Ok(Vc::cell(traced_modules.into_iter().collect()))
+}
+
+/// Ignore non-entry traced reference if not already in tracing mode.
+///
+/// ChunkingType::Traced{TracedMode::Entry}      => target is always traced
+/// ChunkingType::Traced{TracedMode::Transitive} => target only traced if parent is traced
+/// ChunkingType::*                              => target only traced if parent is traced
+fn should_visit_for_tracing(chunking_type: &ChunkingType, parent_traced: bool) -> bool {
+    matches!(
+        chunking_type,
+        ChunkingType::Traced {
+            mode: TracedMode::Entry
+        }
+    ) || parent_traced
+}
+
+#[turbo_tasks::value(transparent, cell = "keyed")]
+pub struct TracedModuleDataIdents(FxHashMap<ResolvedVc<Box<dyn Module>>, ReadRef<AssetIdent>>);
+
+#[turbo_tasks::value(transparent, cell = "keyed")]
+pub struct TracedModuleDataHashes(FxHashMap<ResolvedVc<Box<dyn Module>>, ReadRef<RcStr>>);
+
+#[turbo_tasks::value]
+pub struct TracedModuleData {
+    pub idents: ResolvedVc<TracedModuleDataIdents>,
+    pub hashes: ResolvedVc<TracedModuleDataHashes>,
+}
+
+/// This caches the paths for all modules in the graph so that we don't have to do it once per page.
+#[turbo_tasks::function]
+pub async fn traced_module_data_for_graph(
+    module_graph: Vc<ModuleGraph>,
+    entries_are_traced: bool,
+) -> Result<Vc<TracedModuleData>> {
+    // This function is very similar to traced_modules_for_entries, but doesn't apply the glob and
+    // is executed only once for the whole graph.
+    let module_graph = module_graph.await?;
+    let entries = module_graph.graphs.iter().flat_map(|g| g.entry_modules());
+
+    let mut traced_modules = FxHashSet::default();
+    module_graph.traverse_edges_dfs(
+        entries,
+        &mut (),
+        |parent, target, _| {
+            let Some((parent, ref_data)) = parent else {
+                if entries_are_traced {
+                    traced_modules.insert(target);
+                }
+                return Ok(GraphTraversalAction::Continue);
+            };
+
+            if should_visit_for_tracing(&ref_data.chunking_type, traced_modules.contains(&parent)) {
+                traced_modules.insert(target);
+            };
+            Ok(GraphTraversalAction::Continue)
+        },
+        |_, _, _| Ok(()),
+        true,
+    )?;
+
+    let (idents, hashes): (FxHashMap<_, _>, FxHashMap<_, _>) = traced_modules
+        .into_iter()
+        .map(async |module| {
+            Ok((
+                (module, module.ident().await?),
+                (
+                    module,
+                    module
+                        .source()
+                        .await?
+                        .context("NFT module has no content")?
+                        .content()
+                        .hash(HashAlgorithm::Xxh3Hash128Hex)
+                        .await?,
+                ),
+            ))
+        })
+        .try_join()
+        .await?
+        .into_iter()
+        .unzip();
+
+    Ok(TracedModuleData {
+        idents: ResolvedVc::cell(idents),
+        hashes: ResolvedVc::cell(hashes),
+    }
+    .cell())
+}
+
+/// The globs defined in the next.config.mjs are relative to the project root.
+/// The glob walker in turbopack is somewhat naive so we handle relative path directives first so
+/// traversal doesn't need to consider them and can just traverse 'down' the tree.
+/// The main alternative is to merge glob evaluation with directory traversal which is what the npm
+/// `glob` package does, but this would be a substantial rewrite.
+pub(crate) fn relativize_glob(
+    glob: &str,
+    relative_to: FileSystemPath,
+) -> Result<(&str, FileSystemPath)> {
+    let mut relative_to = relative_to;
+    let mut processed_glob = glob;
+    loop {
+        if let Some(stripped) = processed_glob.strip_prefix("../") {
+            if relative_to.path.is_empty() {
+                bail!(
+                    "glob '{glob}' is invalid, it has a prefix that navigates out of the project \
+                     root"
+                );
+            }
+            relative_to = relative_to.parent();
+            processed_glob = stripped;
+        } else if let Some(stripped) = processed_glob.strip_prefix("./") {
+            processed_glob = stripped;
+        } else {
+            break;
+        }
+    }
+    Ok((processed_glob, relative_to))
+}
+
+#[cfg(test)]
+mod tests {
+    use turbo_tasks::ResolvedVc;
+    use turbo_tasks_backend::{BackendOptions, TurboTasksBackend, noop_backing_storage};
+    use turbo_tasks_fs::{FileSystemPath, NullFileSystem};
+
+    use super::*;
+
+    fn create_test_fs_path(path: &str) -> FileSystemPath {
+        FileSystemPath {
+            fs: ResolvedVc::upcast(NullFileSystem {}.resolved_cell()),
+            path: path.into(),
+        }
+    }
+
+    #[tokio::test(flavor = "multi_thread", worker_threads = 2)]
+    async fn test_relativize_glob_normal_patterns() {
+        let tt = turbo_tasks::TurboTasks::new(TurboTasksBackend::new(
+            BackendOptions::default(),
+            noop_backing_storage(),
+        ));
+        tt.run_once(async {
+            // Test normal glob patterns without relative prefixes
+            let base_path = create_test_fs_path("project/src");
+
+            let (glob, path) = relativize_glob("*.js", base_path.clone()).unwrap();
+            assert_eq!(glob, "*.js");
+            assert_eq!(path.path.as_str(), "project/src");
+
+            let (glob, path) = relativize_glob("components/**/*.tsx", base_path.clone()).unwrap();
+            assert_eq!(glob, "components/**/*.tsx");
+            assert_eq!(path.path.as_str(), "project/src");
+
+            let (glob, path) = relativize_glob("lib/utils.ts", base_path.clone()).unwrap();
+            assert_eq!(glob, "lib/utils.ts");
+            assert_eq!(path.path.as_str(), "project/src");
+            Ok(())
+        })
+        .await
+        .unwrap();
+    }
+
+    #[tokio::test(flavor = "multi_thread", worker_threads = 2)]
+    async fn test_relativize_glob_current_directory_prefix() {
+        let tt = turbo_tasks::TurboTasks::new(TurboTasksBackend::new(
+            BackendOptions::default(),
+            noop_backing_storage(),
+        ));
+        tt.run_once(async {
+            let base_path = create_test_fs_path("project/src");
+
+            // Single ./ prefix
+            let (glob, path) = relativize_glob("./components/*.tsx", base_path.clone()).unwrap();
+            assert_eq!(glob, "components/*.tsx");
+            assert_eq!(path.path.as_str(), "project/src");
+
+            // Multiple ./ prefixes
+            let (glob, path) = relativize_glob("././utils.js", base_path.clone()).unwrap();
+            assert_eq!(glob, "utils.js");
+            assert_eq!(path.path.as_str(), "project/src");
+
+            // ./ with complex glob
+            let (glob, path) = relativize_glob("./lib/**/*.{js,ts}", base_path.clone()).unwrap();
+            assert_eq!(glob, "lib/**/*.{js,ts}");
+            assert_eq!(path.path.as_str(), "project/src");
+            Ok(())
+        })
+        .await
+        .unwrap();
+    }
+
+    #[tokio::test(flavor = "multi_thread", worker_threads = 2)]
+    async fn test_relativize_glob_parent_directory_navigation() {
+        let tt = turbo_tasks::TurboTasks::new(TurboTasksBackend::new(
+            BackendOptions::default(),
+            noop_backing_storage(),
+        ));
+        tt.run_once(async {
+            let base_path = create_test_fs_path("project/src/components");
+
+            // Single ../ prefix
+            let (glob, path) = relativize_glob("../utils/*.js", base_path.clone()).unwrap();
+            assert_eq!(glob, "utils/*.js");
+            assert_eq!(path.path.as_str(), "project/src");
+
+            // Multiple ../ prefixes
+            let (glob, path) = relativize_glob("../../lib/*.ts", base_path.clone()).unwrap();
+            assert_eq!(glob, "lib/*.ts");
+            assert_eq!(path.path.as_str(), "project");
+
+            // Complex navigation with glob
+            let (glob, path) =
+                relativize_glob("../../../external/**/*.json", base_path.clone()).unwrap();
+            assert_eq!(glob, "external/**/*.json");
+            assert_eq!(path.path.as_str(), "");
+            Ok(())
+        })
+        .await
+        .unwrap();
+    }
+
+    #[tokio::test(flavor = "multi_thread", worker_threads = 2)]
+    async fn test_relativize_glob_mixed_prefixes() {
+        let tt = turbo_tasks::TurboTasks::new(TurboTasksBackend::new(
+            BackendOptions::default(),
+            noop_backing_storage(),
+        ));
+        tt.run_once(async {
+            let base_path = create_test_fs_path("project/src/components");
+
+            // ../ followed by ./
+            let (glob, path) = relativize_glob(".././utils/*.js", base_path.clone()).unwrap();
+            assert_eq!(glob, "utils/*.js");
+            assert_eq!(path.path.as_str(), "project/src");
+
+            // ./ followed by ../
+            let (glob, path) = relativize_glob("./../lib/*.ts", base_path.clone()).unwrap();
+            assert_eq!(glob, "lib/*.ts");
+            assert_eq!(path.path.as_str(), "project/src");
+
+            // Multiple mixed prefixes
+            let (glob, path) =
+                relativize_glob("././../.././external/*.json", base_path.clone()).unwrap();
+            assert_eq!(glob, "external/*.json");
+            assert_eq!(path.path.as_str(), "project");
+            Ok(())
+        })
+        .await
+        .unwrap();
+    }
+
+    #[tokio::test(flavor = "multi_thread", worker_threads = 2)]
+    async fn test_relativize_glob_error_navigation_out_of_root() {
+        let tt = turbo_tasks::TurboTasks::new(TurboTasksBackend::new(
+            BackendOptions::default(),
+            noop_backing_storage(),
+        ));
+        tt.run_once(async {
+            // Test navigating out of project root with empty path
+            let empty_path = create_test_fs_path("");
+            let result = relativize_glob("../outside.js", empty_path);
+            assert!(result.is_err());
+            assert!(
+                result
+                    .unwrap_err()
+                    .to_string()
+                    .contains("navigates out of the project root")
+            );
+
+            // Test navigating too far up from a shallow path
+            let shallow_path = create_test_fs_path("project");
+            let result = relativize_glob("../../outside.js", shallow_path);
+            assert!(result.is_err());
+            assert!(
+                result
+                    .unwrap_err()
+                    .to_string()
+                    .contains("navigates out of the project root")
+            );
+
+            // Test multiple ../ that would go out of root
+            let base_path = create_test_fs_path("a/b");
+            let result = relativize_glob("../../../outside.js", base_path);
+            assert!(result.is_err());
+            assert!(
+                result
+                    .unwrap_err()
+                    .to_string()
+                    .contains("navigates out of the project root")
+            );
+            Ok(())
+        })
+        .await
+        .unwrap();
+    }
 }
